candidate_assessment/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.5
django-requestlogs==0.2.2
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
searchtweets-v2==1.0.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tweet-parser==1.13.2
unicodecsv==0.14.1
urllib3==1.26.3
wrapt==1.12.1
